Add dynamic, clickable shortcuts to help panel
Dynmaically add the shortcuts to the help panel
On click, add them to the verbose input, trigger change and hide the help panel.
Replace all ? with a random number.
Use CSS capitalize to make them look nice but not break the case-sensitive regex
Add comments to shortcuts section of translations file