Make the noticed-regex a tooltip and catch more regex-like attempts
Move the noticed-regex warning into the input container so it can
be positioned relative to the input and reduce page height.
Use @extend instead of giving it the .tooltip class as this causes
it to be animated and have it's content changed on hovering over
parts in the results container.
Add a line to "I'm Confused" that explicitly states that this is a
natural language parser for regex - people seem to be confused...!
Move the line about slides into the tutorial section.

Hide the warning tooltip on keydown if it's visible - it was hanging
around too long when re-attempting a description.
Modify the check for regex to look for things like anchors, brackets
and forward slashes as these might also be attempted.